+++ src/_pytask/database_utils.py
@@ -1,12 +1,9 @@
 """Contains utilities for the database."""
 from __future__ import annotations
 
-import hashlib
 from typing import TYPE_CHECKING
 
 from _pytask.dag_utils import node_and_neighbors
-from _pytask.node_protocols import PPathNode
-from _pytask.node_protocols import PTaskWithPath
 from sqlalchemy import Column
 from sqlalchemy import create_engine
 from sqlalchemy import String
@@ -18,10 +15,10 @@ if TYPE_CHECKING:
 
 
 __all__ = [
-    "create_database",
-    "update_states_in_database",
     "BaseTable",
     "DatabaseSession",
+    "create_database",
+    "update_states_in_database",
 ]
 
 
@@ -38,7 +35,6 @@ class State(BaseTable):  # type: ignore[valid-type, misc]
 
     task = Column(String, primary_key=True)
     node = Column(String, primary_key=True)
-    modification_time = Column(String)
     hash_ = Column(String)
 
 
@@ -52,28 +48,14 @@ def create_database(url: str) -> None:
         raise
 
 
-def _create_or_update_state(
-    first_key: str, second_key: str, modification_time: str, hash_: str
-) -> None:
+def _create_or_update_state(first_key: str, second_key: str, hash_: str) -> None:
     """Create or update a state."""
     with DatabaseSession() as session:
         state_in_db = session.get(State, (first_key, second_key))
-
         if not state_in_db:
-            session.add(
-                State(
-                    task=first_key,
-                    node=second_key,
-                    modification_time=modification_time,
-                    hash_=hash_,
-                )
-            )
+            session.add(State(task=first_key, node=second_key, hash_=hash_))
         else:
-            state_in_db.modification_time = (
-                modification_time  # type: ignore[assignment]
-            )
             state_in_db.hash_ = hash_  # type: ignore[assignment]
-
         session.commit()
 
 
@@ -81,17 +63,5 @@ def update_states_in_database(session: Session, task_name: str) -> None:
     """Update the state for each node of a task in the database."""
     for name in node_and_neighbors(session.dag, task_name):
         node = session.dag.nodes[name].get("task") or session.dag.nodes[name]["node"]
-
-        if isinstance(node, PTaskWithPath):
-            modification_time = node.state()
-            hash_ = hashlib.sha256(node.path.read_bytes()).hexdigest()
-        elif isinstance(node, PPathNode):
-            modification_time = node.state()
-            hash_ = ""
-        else:
-            modification_time = ""
-            hash_ = node.state()
-
-        _create_or_update_state(
-            task_name, node.name, modification_time, hash_  # type: ignore[arg-type]
-        )
+        hash_ = node.state()
+        _create_or_update_state(task_name, node.name, hash_)