+++ tests/test_dag.py
@@ -68,6 +68,9 @@ def test_check_if_root_nodes_are_available(tmp_path, runner):
     assert tmp_path.joinpath("in.txt").as_posix() not in result.output
     assert tmp_path.name + "/in.txt" in result.output
 
+    # Test whether reports remove inner tracebacks
+    assert "/_pytask/dag.py" not in result.output
+
 
 @pytest.mark.end_to_end()
 def test_check_if_root_nodes_are_available_w_name(tmp_path, runner):
@@ -212,6 +215,10 @@ def test_error_when_node_state_throws_error(runner, tmp_path):
     result = runner.invoke(cli, [tmp_path.as_posix()])
     assert result.exit_code == ExitCode.DAG_FAILED
     assert "task_example" in result.output
+
+    # Assert that the traceback is hidden.
+    assert "_pytask/nodes.py" not in result.output
+    assert "in state" not in result.output
 
 
 def test_python_nodes_are_unique(tmp_path):